+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -3,15 +3,17 @@ package com.techcourse;
 import com.interface21.webmvc.servlet.HandlerAdapter;
 import com.interface21.webmvc.servlet.HandlerMapping;
 import com.interface21.webmvc.servlet.InitializableHandlerMapping;
-import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.asis.ControllerHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.MethodHandlerAdapter;
+import com.techcourse.handler.HandlerAdapterRegistry;
+import com.techcourse.handler.HandlerDispatcher;
+import com.techcourse.handler.HandlerMappingRegistry;
+import com.techcourse.handler.ManualHandlerMapping;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
 
 public class DispatcherServlet extends HttpServlet {
 
@@ -22,7 +24,7 @@ public class DispatcherServlet extends HttpServlet {
     public DispatcherServlet() {
         handlerMappingRegistry = new HandlerMappingRegistry();
         handlerAdapterRegistry = new HandlerAdapterRegistry();
-        handlerDispatcher = new HandlerDispatcher(handlerAdapterRegistry);
+        handlerDispatcher = new HandlerDispatcher(handlerMappingRegistry, handlerAdapterRegistry);
     }
 
     @Override
@@ -48,38 +50,6 @@ public class DispatcherServlet extends HttpServlet {
     @Override
     protected void service(final HttpServletRequest request, final HttpServletResponse response)
             throws ServletException {
-        try {
-            Object handler = findHandler(request);
-            ModelAndView mav = handlerDispatcher.dispatch(request, response, handler);
-            render(mav, request, response);
-        } catch (IllegalArgumentException e) {
-            handleNotFound(response, e);
-        } catch (Exception e) {
-            handleException(e);
-        }
-    }
-
-    private Object findHandler(final HttpServletRequest request) {
-        return handlerMappingRegistry.getHandler(request)
-                .orElseThrow(() -> new IllegalArgumentException(
-                        "No handler found for " + request.getMethod() + " " + request.getRequestURI()
-                ));
-    }
-
-    private void handleNotFound(final HttpServletResponse response, final IllegalArgumentException e) 
-            throws ServletException {
-        try {
-            response.sendError(HttpServletResponse.SC_NOT_FOUND, e.getMessage());
-        } catch (IOException ex) {
-            throw new ServletException(ex);
-        }
-    }
-
-    private void handleException(final Exception e) throws ServletException {
-        throw new ServletException(e.getMessage());
-    }
-
-    protected void render(ModelAndView mav, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        mav.getView().render(mav.getModel(), request, response);
+        handlerDispatcher.execute(request, response);
     }
 }
