+++ app/src/main/java/com/techcourse/handler/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.handler;
 
 import com.interface21.webmvc.servlet.HandlerAdapter;
 import java.util.ArrayList;
